ruby: fix double-free in the vizkit3d-to-ruby adapter stuff
The issue was that the qruby_bridge object was both deleted by
the TypelibToQVariant objects *and* Qt. When the application quits,
if the former was first then everything was fine. However, if the
QApplication object was deleted *before* the TypelibToQVariant
was, then boom.

Make the qruby_bridge a child of the qApp only during the time
needed to get a hold of it. Remove the relationship afterwards.