plugins: always use the plugin factory to create vizkit plugins
The reason is that we want to be able to create multiple times
the same plugin. The Qt plugin architecture is based on the
assumption that the plugin *is* a factory, i.e. that it can
be loaded only once per library, and then used to create objects.

This means that, before this commit, doing createPlugin('lib', 'plugin')
twice was getting us the same plugin instead of creating it twice.