normalize all callback functions to the UnboundMethod API
All callback functions are now objects that support

  #call(receiver, *args, &block)

and

  #bind(receiver) => value that supports #call(*args, &block)

It is therefore now supported to give a block to the #register* methods.
Users of the callback methods only have to call #call and/or first #bind
and #call to get their functional callbacks. Moreover, blocks are passed
through the #call interface as the last argument (1.8 does not support
passing blocks to blocks)