ruby TaskProxy: speed up pinging for situations where multiple task go offline
ping is internal implement by a remote message call which goes into a
timeout after the task went offline. Normally the timeout is set around
3 seconds which means that if 10 task go offline the ruby script is blocking
for around 30 seconds. Therefor if one task is going offline all tasks are
checked against the nameserver and all tasks which are missing are
regarded to be offline as well.