vizkit plugin: bugfix related to m types
The plugin loader is not able to normalize types if orocos is
uninitialized which is unfortunately the case when the plugins are
loaded. Therefore the new default behavior is that all m types
are normalized to their opaque type and all widgets must be registered
as none m type.

Example:
RigidBodyState_m must only be registered as RigidBodyState