ruby: fix the tree modeller to require much less dispatching
Currently, a single huge method update_object was called for
every object that needed some generation. This method has been
split into various update_AAA methods, for each of the display-able
types that the tree modeller understands.

update_object is for backward compatibility, and because some cases
are not covered (yet) by the various update_* methods.

However, we now call specific update_AAA methods when the context
allows to know what to call. This makes the whole class a lot less
dependent on the specific object types, which gives us more flexibility
(i.e. one assumes that a value that each_output_port yields conforms
to the output port API).

In general, it allows the usage of rock-display with ROS nodes under
rock