rock-transformer: add SDF support
This commit adds the ability to load a SDF file in rock-transformer
(or to load a SDF file from the transformer's configuration file,
which is the most likely use case). If in design mode, it will show
a joint-setting UI. Otherwise, it assumes that the joint transformations
are produced dynamically by the system.